refactor: 파일 구조 변경 및 라벨 수정
