+++ Diary/Network/HttpMethod.swift
@@ -1,0 +1,11 @@
+//
+//  HttpMethod.swift
+//  Diary
+//
+//  Created by Christy, vetto on 2023/05/09.
+//
+
+enum HttpMethod: String {
+    case get = "GET"
+    case post = "POST"
+}
